Consolidate metadata into `metaLsx` class
