+++ src/library/bg3/index.ts
@@ -4,6 +4,5 @@
 
 export * from './convertToHandle';
 export * from './Version';
-export * from './getMetadata';
 export * from './buildLocalizationReference';
 export * from './meta';
