+++ src/library/index.ts
@@ -2,4 +2,4 @@
 // LIBRARY
 // =======
 
-export * as bg3 from './bg3/index';
+export * as bg3 from './bg3';
